requirements.txt
openai
python-dotenv
weave
instructor